game player logic, game ui and gfx changes
Initial game player logic has been implemented, but it is not used
currently. The single player game menu and the in-game menu has been
improved.

All real and artificial engine limits are now easily accessible in
genie/include/genie/limits.h. This makes it easier for the engine to
keep track of certain limits as well for tools that might need them.

The graphics subsystem now only accepts integer coordinates for text
because floating point numbers may resulted in some cases in blurry
unreadable text.

The user interface now also accepts a display function pointer for
menu-specific objects (e.g. in-game objects). Also fixed the scenario
builder button coordinates.

+++ genie/gfx.h
@@ -22,7 +22,7 @@ void genie_gfx_setup_ortho(GLdouble width, GLdouble height);
 
 void genie_gfx_clear_screen(GLclampf red, GLclampf green, GLclampf blue, GLclampf alpha);
 
-void genie_gfx_draw_text(GLfloat x, GLfloat y, const char *str);
+void genie_gfx_draw_text(int x, int y, const char *str);
 
 #define GENIE_HA_LEFT 0
 #define GENIE_HA_CENTER 1
